+++ packages/app/components/PlotDrawer/index.tsx
@@ -1,0 +1,216 @@
+import React, { useCallback, useEffect } from "react";
+import { Animated, Dimensions, StyleSheet, View } from "react-native";
+import { Queue, Stack } from "react-native-spacing-system";
+import { BlurView } from "expo-blur";
+import dateformat from "dateformat";
+import { Forest, Plot } from "@ong-forestry/schema";
+import { MapScreenModes, DrawerStates } from "../../constants";
+import useAppSelector from "../../hooks/useAppSelector";
+import { locallyDeleteTree, deselectTree } from "../../redux/slices/treeSlice";
+import DrawerButton from "../DrawerButton";
+import { Text, TextVariants } from "../Themed";
+import useAppDispatch from "../../hooks/useAppDispatch";
+import DataEntryForm from "./DataEntryForm";
+
+interface PlotDrawerProps {
+  mode: MapScreenModes;
+  drawerState: DrawerStates;
+  plot?: Plot;
+  forest?: Forest;
+  setDrawerHeight: (height: number) => void;
+  openVisualizationModal: () => void;
+  beginPlotting: () => void;
+  endPlotting: () => void;
+  expandDrawer: () => void;
+  minimizeDrawer: () => void;
+}
+
+export const PlotDrawer: React.FC<PlotDrawerProps> = ({
+  mode,
+  drawerState,
+  plot,
+  setDrawerHeight,
+  beginPlotting,
+  expandDrawer,
+  minimizeDrawer,
+}) => {
+  useEffect(() => {
+    return function cleanup() {
+      setDrawerHeight(0);
+    };
+  }, []);
+
+  const dispatch = useAppDispatch();
+
+  const {
+    all,
+    selected: selectedTreeTag,
+    indices: { byPlots },
+  } = useAppSelector((state) => state.trees);
+  const selected = selectedTreeTag ? all[selectedTreeTag] : undefined;
+
+  const setStyle = useCallback(() => {
+    switch (drawerState) {
+      case DrawerStates.Closed:
+        return styles.containerClosed;
+      case DrawerStates.Minimized:
+        return styles.containerMinimized;
+      case DrawerStates.Expanded:
+        return styles.containerExpanded;
+    }
+  }, [drawerState]);
+
+  const computePlotLastUpdatedDate = useCallback(
+    (plotNumber: string) => {
+      const plotTrees = byPlots[plotNumber];
+      let latestCensus: Date | undefined;
+      for (const treeTag of plotTrees) {
+        const { updatedAt } = all[treeTag];
+        if (!latestCensus || updatedAt > latestCensus) {
+          latestCensus = updatedAt;
+        }
+      }
+      return latestCensus;
+    },
+    [byPlots]
+  );
+
+  if (drawerState === DrawerStates.Closed) {
+    return null;
+  }
+
+  return (
+    <Animated.View
+      style={{ ...styles.container, ...setStyle() }}
+      onLayout={(e) => {
+        setDrawerHeight(e.nativeEvent.layout.height);
+      }}
+    >
+      <BlurView style={styles.blurContainer} intensity={40}>
+        {mode === MapScreenModes.Select && !!plot && (
+          <View style={styles.header}>
+            <View style={{ flexDirection: "row", alignItems: "center" }}>
+              <Text variant={TextVariants.H2}>Plot #{plot.number}</Text>
+              <Queue size={36}></Queue>
+              <>
+                {(() => {
+                  const lastUpdated = computePlotLastUpdatedDate(plot.number);
+                  return (
+                    <Text variant={TextVariants.Body}>
+                      {lastUpdated
+                        ? `Last censused on ${dateformat(
+                            lastUpdated,
+                            "mmm dS, yyyy"
+                          )}`
+                        : "Never censused"}
+                    </Text>
+                  );
+                })()}
+              </>
+            </View>
+            {drawerState === "MINIMIZED" && (
+              <DrawerButton onPress={beginPlotting}>Add Trees</DrawerButton>
+            )}
+          </View>
+        )}
+        {mode === MapScreenModes.Explore && (
+          <View style={[styles.header]}>
+            <View style={{ flexDirection: "row", alignItems: "center" }}>
+              <Text variant={TextVariants.H2}>O-Farm</Text>
+              <Queue size={36}></Queue>
+              <Text variant={TextVariants.Body}>Select any plot to begin</Text>
+            </View>
+          </View>
+        )}
+        {mode === MapScreenModes.Plot && !!plot && (
+          <>
+            <View style={styles.header}>
+              {drawerState === "MINIMIZED" && !selected && (
+                <View style={{ flexDirection: "row", alignItems: "center" }}>
+                  <Text variant={TextVariants.H2}>Plot #{plot.number}</Text>
+                  <Queue size={36}></Queue>
+                  <Text variant={TextVariants.Body}>
+                    Tap anywhere to plot a new tree
+                  </Text>
+                </View>
+              )}
+              {drawerState === "MINIMIZED" && !!selected && (
+                <>
+                  <View style={{ flexDirection: "row", alignItems: "center" }}>
+                    <Text variant={TextVariants.H2}>Tree #{selected.tag}</Text>
+                    <Queue size={36}></Queue>
+                    <Text variant={TextVariants.Body}>
+                      Last census on{" "}
+                      {dateformat(
+                        selected.updatedAt,
+                        'mmm dS, yyyy "at" h:MM TT'
+                      )}
+                    </Text>
+                  </View>
+                  <DrawerButton onPress={expandDrawer}>Edit</DrawerButton>
+                </>
+              )}
+              {drawerState === "EXPANDED" && !!selected && (
+                <>
+                  <Text variant={TextVariants.H2}>New tree</Text>
+                  <Text variant={TextVariants.Body}>
+                    This is the {byPlots[plot.number].size + 1}th tree in the
+                    plot.
+                  </Text>
+                </>
+              )}
+            </View>
+
+            {drawerState === "EXPANDED" && !!selected && (
+              <>
+                <Stack size={24}></Stack>
+                <View style={styles.content}>
+                  <DataEntryForm
+                    cancel={() => {
+                      dispatch(locallyDeleteTree(selected.tag));
+                      dispatch(deselectTree());
+                      minimizeDrawer();
+                    }}
+                    finish={minimizeDrawer}
+                    style={{ flex: 1 }}
+                  />
+                </View>
+              </>
+            )}
+          </>
+        )}
+      </BlurView>
+    </Animated.View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    position: "absolute",
+    bottom: 0,
+    zIndex: 2,
+    width: Dimensions.get("window").width,
+    backgroundColor: "rgba(255, 255, 255, 0.4)",
+    borderTopLeftRadius: 24,
+    borderTopRightRadius: 24,
+    overflow: "hidden",
+  },
+  blurContainer: {
+    width: "100%",
+    height: "100%",
+    padding: 24,
+  },
+  containerClosed: {
+    display: "none",
+  },
+  containerMinimized: {},
+  containerExpanded: {},
+  header: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  content: {
+    // height: 512,
+  },
+});
