+++ packages/app/package.json
@@ -9,7 +9,9 @@
     "ios": "expo start --ios",
     "web": "expo start --web",
     "eject": "expo eject",
-    "test": "jest --watchAll"
+    "test": "jest --watchAll",
+    "lint": "eslint **/**/*.{ts,tsx}"
+
   },
   "jest": {
     "preset": "jest-expo"
@@ -22,8 +24,6 @@
     "@react-navigation/native": "^6.0.2",
     "@react-navigation/native-stack": "^6.1.0",
     "@reduxjs/toolkit": "^1.8.0",
-    "@types/color": "3.0.0",
-    "@types/dateformat": "^5.0.0",
     "axios": "^0.26.0",
     "color": "3.0.0",
     "dateformat": "^5.0.3",
@@ -56,10 +56,17 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
+    "@types/color": "^3.0.3",
+    "@types/dateformat": "^5.0.0",
     "@types/lodash": "^4.14.179",
     "@types/react": "~17.0.21",
     "@types/react-native": "~0.64.12",
     "@types/utm": "^1.1.1",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
+    "eslint": "^8.12.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-native": "^4.0.0",
     "jest": "^26.6.3",
     "jest-expo": "~44.0.1",
     "react-test-renderer": "17.0.1",
