+++ packages/server/package.json
@@ -11,12 +11,6 @@
   },
   "dependencies": {
     "@ong-forestry/schema": "1.0.0",
-    "@types/bcrypt": "^5.0.0",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/morgan": "^1.9.3",
-    "@types/passport-jwt": "^3.0.6",
-    "@types/passport-local": "^1.0.34",
     "bcrypt": "^5.0.1",
     "body-parser": "^1.19.2",
     "cors": "^2.8.5",
@@ -42,6 +36,12 @@
     "uuid4": "^2.0.2"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.0",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/morgan": "^1.9.3",
+    "@types/passport-jwt": "^3.0.6",
+    "@types/passport-local": "^1.0.34",
     "ts-node-dev": "^1.1.8"
   }
 }
